% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDistributionFunction.R
\name{getDistributionFunction}
\alias{getDistributionFunction}
\title{Getting distribution functions}
\usage{
getDistributionFunction(type, dist, ...)
}
\arguments{
\item{type}{Character, typically either 'r', 'q', 'p', or  'd'.}

\item{dist}{Character, typically something like 'norm', 'gamma', etc.}

\item{...}{Currently ignored.}
}
\value{
Function, the first function in the search path that matches the name \code{paste0(type, dist)}.
}
\description{
Determines if the distribution functions are available. This is intended for internal use only.
}
\details{
It is determined that \code{paste0(type, dist)} is a function and returns that function.  The nature of the returned function is not verified.
}
\examples{

fun <- getDistributionFunction(type="q",dist="norm")

}
