% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.windAC.R
\name{print.windAC}
\alias{print.windAC}
\title{Print \code{windAC} object}
\usage{
\method{print}{windAC}(x, ...)
}
\arguments{
\item{x}{A \code{windAC} object.}

\item{...}{Currently ignored.}
}
\value{
Print for windAC object
}
\description{
Print \code{windAC} object.
}
\details{
see \code{\link{calcAC}}
}
