% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turbineSpatial.R
\docType{data}
\name{turbineSpatial}
\alias{turbineSpatial}
\title{Data sets of spatial area searched at a wind farm}
\format{The \code{turbinePoints} element is a spatial data frame with:
\describe{
  \item{\code{turbName}}{The name of the turbine for the search plot.}
}}
\usage{
data(turbineSpatial)
}
\description{
List containing two spatial data frames
}
\details{
A list of data frames, with two elements named \code{turbinePlots} and \code{turbinePoints}. These are fictious data for example purposes only.
}
\keyword{datasets}
