% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windAC.R
\docType{package}
\name{windAC-package}
\alias{windAC-package}
\title{windAC: Area Correction Methods}
\description{
Post-construction fatality monitoring studies at wind facilities are based on data from searches for bird and bat carcasses in plots beneath turbines. Bird and bat carcasses can fall outside of the search plot. Bird and bat carcasses from wind turbines often fall outside of the searched area. To compensate, area correction (AC) estimations are calculated to estimate the percentage of fatalities that fall within the searched area versus those that fall outside of it. This package provides two likelihood based methods and one physics based method (Hull and Muir (2010) \doi{10.1080/14486563.2010.9725253}, Huso and Dalthorp (2014) \doi{10.1002/jwmg.663}) to estimate the carcass fall distribution. There are also functions for calculating the proportion of area searched within one unit annuli, log logistic distribution functions, and truncated distribution functions.
}
\author{
\strong{Maintainer}: Daniel Riser-Espinoza \email{despinoza@west-inc.com}

Authors:
\itemize{
  \item Paul Rabie
}

Other contributors:
\itemize{
  \item Trent McDonald [contributor]
  \item Western EcoSystems Technology, Inc. [copyright holder, funder]
}

}
\keyword{internal}
