% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turbineSpatial.R
\docType{data}
\name{turbineSpatial}
\alias{turbineSpatial}
\title{Data sets of spatial area searched at a wind farm}
\format{
An object of class \code{list} of length 2.
}
\usage{
data(turbineSpatial)
}
\description{
List containing two spatial data frames
}
\details{
A list of data frames, with two elements named \code{turbinePlots} and \code{turbinePoints}. These are fictious data for example purposes only. The \code{turbinePlots} element is a spatial polygon data frame with the polygons being the area searched round the turbines. The \code{turbinePoints} element is a spatial points data frame with the turbine locations.
}
\keyword{datasets}
